Organize the classes into seperate files
